+++ share/dotfiles/.config/nwg-dock-hyprland/launch.sh
@@ -4,5 +4,6 @@
 # /____/\___/\__/_/\_\ 
 #                      
 
-# Add exec-once = ~/.config/nwg-dock-hyprland/launch.sh to ~/.config/hypr/conf/custom.conf
-nwg-dock-hyprland -i 32 -w 5 -mb 10 -ml 10 -mr 10 -x -c "rofi -show drun"
+if [ -f $HOME/.config/ml4w/settings/nwg-dock-hyprland.sh ] ;then
+    nwg-dock-hyprland -i 32 -w 5 -mb 10 -ml 10 -mr 10 -x -c "rofi -show drun"
+fi